django-microservice/app/serializers.py
from rest_framework import serializers

from .models import Recipe, RecipeComment


class RecipeSerializer(serializers.ModelSerializer):
    class Meta:
        model = Recipe
        fields = '__all__'

class RecipeCommentSerializer(serializers.ModelSerializer):
    class Meta:
        model = RecipeComment
        fields = '__all__'