django-microservice/requirements.txt
Django==3.2
djangorestframework==3.14.0
pika
psycopg2==2.8.6